requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
astroid==2.3.1
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
deepdiff==4.0.7
discord.py==1.2.3
entrypoints==0.3
flake8==3.7.8
idna==2.8
isort==4.3.21
jsonpickle==1.2
lazy-object-proxy==1.4.2
logmatic-python==0.1.7
mccabe==0.6.1
multidict==4.5.2
ordered-set==3.1.1
pycares==3.0.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.4.2
python-dateutil==2.8.0
python-json-logger==0.1.11
PyYAML==5.1.2
six==1.12.0
typed-ast==1.4.0
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
